LayoutTests/fast/misc/testProcessIncomingSyncMessagesWhenWaitingForSyncReply-expected.txt

Test for the SendSyncOption::ProcessIncomingSyncMessagesWhenWaitingForSyncReply sendSync flag

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.testProcessIncomingSyncMessagesWhenWaitingForSyncReply() is true
PASS successfullyParsed is true

TEST COMPLETE
